requirements.txt
aiohttp
aiosqlite
bs4
jinja2
pillow
requests
